requirements.txt
# Зависимости SnapCrawler (с пояснениями)

# Работа с YAML-конфигом
pyyaml>=6.0.1

# Парсинг HTML (поиск <img>, извлечение ссылок)
beautifulsoup4>=4.12.3
lxml>=5.2.2                # Быстрый XML/HTML-парсер для bs4

# Асинхронные HTTP-запросы
aiohttp>=3.9.5             # Основной клиент для загрузки изображений
httpx>=0.27.0              # (опционально) резерв/утилиты; сейчас не используется

# Мимикрия клиента
fake-useragent>=1.5.1      # (опционально) ротация User-Agent; сейчас не используется

# RSS/Atom (опционально)
feedparser>=6.0.11         # (опционально) парсинг RSS; сейчас не используется

# Работа с изображениями
Pillow>=10.4.0             # Предобработка, конвертация JPEG
numpy>=2.0.1               # Быстрые операции над массивами (вариации, преобразования)
imagehash>=4.3.1           # pHash для дедупликации

# Нейросетевой инференс (опционально)
onnxruntime>=1.18.0        # CPU-инференс ONNX‑модели для фильтра фото/не‑фото
scikit-image>=0.24.0       # (опционально) SSIM; сейчас не используется

# Планировщик задач/паузы (опционально)
APScheduler>=3.10.4        # (опционально) сейчас не используется

# Прогресс-бары и вспомогательные утилиты
tqdm>=4.66.4               # (опционально) сейчас не используется
